+++ packages/ricos-editor/web/src/RicosEditor.spec.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable max-len */
 import React from 'react';
-import { RicosEditor, RicosEditorProps } from './index';
+import { RicosEditor, RicosEditorProps, DraftEditorSettings } from './index';
 import { RichContentEditor } from 'wix-rich-content-editor';
 import introState from '../../../../e2e/tests/fixtures/intro.json';
 import { pluginHashtag } from '../../../plugin-hashtag/web/src/editor';
@@ -142,6 +142,19 @@ describe('RicosEditor', () => {
     const rceProps_noTheme = JSON.stringify({ ...rceProps, theme: {} });
     const rcePropsWrapped_noTheme = JSON.stringify({ ...rcePropsWrapped, theme: {} });
     expect(rceProps_noTheme).toStrictEqual(rcePropsWrapped_noTheme);
+  });
+  it('should only accept valid Draft-js editor props', () => {
+    const draftEditorSettings: DraftEditorSettings & { notADraftSetting: boolean } = {
+      tabIndex: -1,
+      spellCheck: true,
+      stripPastedStyles: false,
+      notADraftSetting: false,
+    };
+    const rceProps = getRCE({ draftEditorSettings }).props();
+    expect(rceProps).toHaveProperty('theme');
+    expect(rceProps.tabIndex).toEqual(-1);
+    expect(rceProps.spellCheck).toEqual(true);
+    expect(rceProps).not.toHaveProperty('notADraftSetting');
   });
   describe('Modal API', () => {
     it('should pass openModal & closeModal to helpers', () => {